refactor: 💡 makes macOS format output an Objective-C file
Unfortunately, there seems to be no way to directly generate a macOS
.clr file in JS as it's an undocumented binary format. Instead, the
approach seems to be to write an Objective-C app, which can be compiled
with XCode and, when run, generates a color list and writes it out as a
.clr file. This commit takes steps towards that by updating the
StyleDictionary config and macOS format to output a (work in progress)
temporary Objective-C file. Similar to what we do with TypeScript, we
should eventually be able to have a follow-on task that compiles that
Objective-C code and runs it to generate the .clr file.