feat(colors): adjusts colors to improve contrast ratios
Tweaks a number of color values and also the assignments to color purposes in the various color
schemes, in order to conform to the new, stricter contrast ratio requirements. Note, the naming of
the "Wipro extended" shades of grey was getting a bit out of hand, so a new scheme using the
lightness percentage (from the HSL representation) has been introduced. For now, the old names have
been preserved and alias the new ones (to avoid breaking changes), but these names should be
considered deprecated and will be removed in the next major release.

30