build: swtiches to gulp-unzip to avoid security warnings
We were using gulp-decompress to unzip the macOS ColorTools we download as part of the build, but
one of its dependencies has a critical security issue reported by npm audit. gulp-unzip, which is an
alternative does not appear to have any known security issues, so this change switches to that.
Unlike decompress, the extracted utils are not executable though, so they now need to be chmodded
too. This change also updates various other dependencies to their latest versions.